functions/collector/requirements.txt
functions-framework==3.*
google-cloud-secret-manager==2.*
google-cloud-storage==2.*
LivePopularTimes
requests
pytz
